# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Bug Bounty Toolkit application.",
        "title": "Toolkit API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1.0.0"
    },
    "host": "localhost:8778",
    "basePath": "/api",
    "paths": {
...
        "/scope-rules": {
            "get": {
                "description": "Retrieves all scope rules for a specified target.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scope Rules"
                ],
                "summary": "Get scope rules for a target",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the target to fetch rules for",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of scope rules for the target",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ScopeRule"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or missing target_id",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Target not found (if target_id does not exist)",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new scope rule (in-scope or out-of-scope) to a specified target.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scope Rules"
                ],
                "summary": "Add a scope rule to a target",
                "parameters": [
                    {
                        "description": "Scope rule to add. 'target_id' is required. 'id' and 'is_wildcard' are ignored/auto-set by backend.",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScopeRule"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Scope rule created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.ScopeRule"
                        }
                    },
                    "400": {
                        "description": "Invalid request body, missing required fields, or invalid target_id/item_type",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Target not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Duplicate scope rule",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scope-rules/{ruleId}": {
            "get": {
                "description": "Handles GET, PUT, DELETE for a specific scope rule by its ID.",
                "tags": [
                    "Scope Rules"
                ],
                "summary": "Scope Rule item operations",
                "responses": {}
            },
            "put": {
                "description": "Handles GET, PUT, DELETE for a specific scope rule by its ID.",
                "tags": [
                    "Scope Rules"
                ],
                "summary": "Scope Rule item operations",
                "responses": {}
            },
            "delete": {
                "description": "Deletes a scope rule by its numeric ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scope Rules"
                ],
                "summary": "Delete a specific scope rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the scope rule to delete",
                        "name": "ruleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Scope rule deleted successfully"
                    },
                    "400": {
                        "description": "Invalid rule ID format",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Scope rule not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message describing the issue"
                }
            }
        },
...
        "models.ScopeRule": {
            "type": "object",
            "required": [
                "is_in_scope",
                "item_type",
                "pattern"
            ],
            "properties": {
                "description": {
                    "description": "Optional description for the scope rule.",
                    "type": "string",
                    "example": "Main application domain"
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 101
                },
                "is_in_scope": {
                    "description": "True if the pattern is in scope, false if out of scope.",
                    "type": "boolean",
                    "example": true
                },
                "is_wildcard": {
                    "description": "True if the pattern contains a wildcard, auto-detected by backend.",
                    "type": "boolean",
                    "example": true
                },
                "item_type": {
                    "description": "Type of scope item.",
                    "type": "string",
                    "example": "domain"
                },
                "pattern": {
                    "description": "The scope pattern.",
                    "type": "string",
                    "example": "*.example.com"
                },
                "target_id": {
                    "description": "ID of the target this rule belongs to.",
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                }
            }
        },
...
    }
}